+++ filer/static/filer/js/addons/popup_handling.js
@@ -19,10 +19,11 @@ if (django.jQuery) {
         document.location.reload();
         win.close();
     };
-    window.dismissRelatedImageLookupPopup = function (win, chosenId, chosenThumbnailUrl, chosenDescriptionTxt) {
+    window.dismissRelatedImageLookupPopup = function (win, chosenId, chosenThumbnailUrl, chosenDescriptionTxt, chosenAdminChangeUrl) {
         var id = windowname_to_id(win.name);
         var lookup = $('#' + id);
         var container = lookup.closest('.filerFile');
+        var edit = container.find('.edit');
         var image = container.find('.thumbnail_img');
         var descriptionText = container.find('.description_text');
         var clearer = container.find('.filerClearer');
@@ -32,11 +33,15 @@ if (django.jQuery) {
 
         element.val(chosenId);
         element.closest('.js-filer-dropzone').addClass('js-object-attached');
-        image.attr('src', chosenThumbnailUrl).removeClass('hidden');
-        image.removeAttr('srcset');  // would be nicer, but much more complicate to also replace 'srcset'
+        if (chosenThumbnailUrl) {
+            image.attr('src', chosenThumbnailUrl).removeClass('hidden');
+            image.removeAttr('srcset');  // would be nicer, but much more complicate to also replace 'srcset'
+        }
         descriptionText.text(chosenDescriptionTxt);
         clearer.removeClass('hidden');
         lookup.addClass('related-lookup-change');
+        edit.addClass('related-lookup-change');
+        if (chosenAdminChangeUrl) edit.attr('href', chosenAdminChangeUrl + '?_edit_from_widget=1');
         dropzoneMessage.addClass('hidden');
 
         if (oldId !== chosenId) {